requirements.txt
PyYAML
requests
bs4
django-bootstrap4